Add files via upload
If you have any question just put it here below, I'll be happy to help you

We do not need for every index iterate over all array, what we need is just take the sum of all array and then just subtruct for every new index the value sits on that index and update the length of new array.

left is just sum off all values to the left of the index, and right is the opposite. l (L) is the number of values combined into left, r is the number of values combined into right.

In every iteration, we subtruct the current value from right and add it to left. Also we subtruct 1 from r (r -= 1) and add it to l, since right is getting shorter and opposite happens to left.